requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
filelock==3.16.1
idna==3.10
lxml==5.3.0
oauthlib==3.2.2
PySocks==1.7.1
python-decouple==3.8
requests==2.32.3
requests-oauthlib==1.3.1
scraper==0.1.0
snscrape==0.7.0.20230622
soupsieve==2.6
sqlparse==0.5.3
tweepy==4.14.0
tzdata==2024.2
urllib3==2.3.0
